#42: refactor types and utils
