+++ utils/utils.go
@@ -2,57 +2,11 @@ package utils
 
 import (
 	"crypto/ecdsa"
-	"encoding/json"
 	"errors"
-	"log"
-	"math/big"
-	"os"
-	"path/filepath"
-	"regexp"
-
 	gethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
-	"gopkg.in/yaml.v3"
+	"math/big"
 )
-
-func ReadFile(path string) ([]byte, error) {
-	b, err := os.ReadFile(filepath.Clean(path))
-	if err != nil {
-		return nil, err
-	}
-	return b, nil
-}
-
-func ReadYamlConfig(path string, o interface{}) error {
-	if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
-		log.Fatal("Path ", path, " does not exist")
-	}
-	b, err := ReadFile(path)
-	if err != nil {
-		return err
-	}
-
-	err = yaml.Unmarshal(b, o)
-	if err != nil {
-		log.Fatalf("unable to parse file with error %#v", err)
-	}
-
-	return nil
-}
-
-func ReadJsonConfig(path string, o interface{}) error {
-	b, err := ReadFile(path)
-	if err != nil {
-		return err
-	}
-
-	err = json.Unmarshal(b, o)
-	if err != nil {
-		log.Fatalf("unable to parse file with error %#v", err)
-	}
-
-	return nil
-}
 
 func EcdsaPrivateKeyToAddress(privateKey *ecdsa.PrivateKey) (gethcommon.Address, error) {
 	publicKey := privateKey.Public()
@@ -70,9 +24,4 @@ func RoundUpDivideBig(a, b *big.Int) *big.Int {
 	a.Sub(a, one)
 	res.Div(a, b)
 	return res
-}
-
-func IsValidEthereumAddress(address string) bool {
-	re := regexp.MustCompile("^0x[0-9a-fA-F]{40}$")
-	return re.MatchString(address)
 }
